Extract a singleton QuickTriggerable comparator to simplify the class
